Split expiration_badge.dart expiration_text.dart, display_checklist.dart, edit_checklist.dart
