+++ test/presentation/pages/task_info/widgets/display_checklist_test.dart
@@ -2,11 +2,10 @@ import 'package:tasky/core/ilist.dart';
 import 'package:tasky/domain/entities/checklist.dart';
 import 'package:tasky/domain/values/task_values.dart';
 import 'package:tasky/domain/values/unique_id.dart';
-import 'package:tasky/presentation/pages/task_form/misc/checklist_primitive.dart';
-import 'package:tasky/presentation/widgets/checklist_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_feather_icons/flutter_feather_icons.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:tasky/presentation/pages/task_info/widgets/display_checklist.dart';
 
 void main() {
   group("DisplayChecklist test", () {
@@ -225,91 +224,6 @@ void main() {
       final percent =
           widget.createState().getChecklistProgress(widget.checklist.items);
       expect(percent, equals(1 / 3));
-    });
-  });
-
-  group("EditChecklist test", () {
-    testWidgets("show widget correctly", (tester) async {
-      await tester.pumpWidget(
-        MaterialApp(
-          home: Scaffold(
-            body: EditChecklist(
-              primitive: ChecklistPrimitive(
-                title: "Checklist",
-                items: IList.from(
-                  [
-                    ItemText("Item 1"),
-                    ItemText("Item 2"),
-                    ItemText("Item 3"),
-                  ],
-                ),
-              ),
-            ),
-          ),
-        ),
-      );
-
-      expect(find.text("Checklist"), findsOneWidget);
-      expect(find.text("Item 1"), findsOneWidget);
-      expect(find.text("Item 2"), findsOneWidget);
-      expect(find.text("Item 3"), findsOneWidget);
-      expect(find.byType(Checkbox), findsNWidgets(3));
-    });
-
-    testWidgets("tap call edit correctly", (tester) async {
-      bool edit = false;
-      await tester.pumpWidget(
-        MaterialApp(
-          home: Scaffold(
-            body: EditChecklist(
-              onTap: () => edit = true,
-              primitive: ChecklistPrimitive(
-                title: "Checklist",
-                items: IList.from(
-                  [
-                    ItemText("Item 1"),
-                    ItemText("Item 2"),
-                    ItemText("Item 3"),
-                  ],
-                ),
-              ),
-            ),
-          ),
-        ),
-      );
-
-      expect(edit, isFalse);
-      await tester.tap(find.byType(Card));
-      await tester.pumpAndSettle();
-      expect(edit, isTrue);
-    });
-
-    testWidgets("tap on delete icon call remove correctly", (tester) async {
-      bool remove = false;
-      await tester.pumpWidget(
-        MaterialApp(
-          home: Scaffold(
-            body: EditChecklist(
-              onRemove: () => remove = true,
-              primitive: ChecklistPrimitive(
-                title: "Checklist",
-                items: IList.from(
-                  [
-                    ItemText("Item 1"),
-                    ItemText("Item 2"),
-                    ItemText("Item 3"),
-                  ],
-                ),
-              ),
-            ),
-          ),
-        ),
-      );
-
-      expect(remove, isFalse);
-      await tester.tap(find.byIcon(Icons.delete));
-      await tester.pumpAndSettle();
-      expect(remove, isTrue);
     });
   });
 }
